1.3.3

+++ src/main/java/be/isach/ultracosmetics/cosmetics/morphs/MorphCreeper.java
@@ -27,7 +27,7 @@ import java.util.UUID;
  */
 public class MorphCreeper extends Morph {
 
-    private double charge = 0;
+    private int charge = 0;
 
     public MorphCreeper(UUID owner) {
         super(DisguiseType.CREEPER, Material.SULPHUR, (byte) 0, "Creeper", "ultracosmetics.morphs.creeper", owner, MorphType.CREEPER, "&7&oNice housssssssse you got! Sssssssh");
